requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
bottleneck=1.3.7=py311hcb2cf0a_0
brotli-python=1.1.0=py311hb755f60_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.7.22=hbcca054_0
certifi=2023.7.22=pyhd8ed1ab_0
charset-normalizer=3.2.0=pyhd8ed1ab_0
h5py=3.7.0=py311h021c08c_0
hdf5=1.10.6=h3ffc7dd_1
idna=3.4=pyhd8ed1ab_0
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=17_linux64_openblas
libcblas=3.9.0=17_linux64_openblas
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.1.0=he5830b7_0
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
liblapack=3.9.0=17_linux64_openblas
libopenblas=0.3.23=pthreads_h80387f5_0
libstdcxx-ng=13.1.0=hfd8a6a1_0
libuuid=1.41.5=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
ncurses=6.4=h6a678d5_0
numpy=1.25.2=py311h64a7726_0
openssl=3.1.2=hd590300_0
packaging=23.1=pyhd8ed1ab_0
physics-tenpy=0.10.0=py311h4dd048b_1
pip=23.2.1=py311h06a4308_0
platformdirs=3.10.0=pyhd8ed1ab_0
pooch=1.7.0=pyha770c72_3
pysocks=1.7.1=pyha2e5f31_6
python=3.11.4=h955ad1f_0
python_abi=3.11=2_cp311
pyyaml=6.0.1=py311h459d7ec_0
readline=8.2=h5eee18b_0
requests=2.31.0=pyhd8ed1ab_0
scipy=1.11.2=py311h64a7726_0
setuptools=68.0.0=py311h06a4308_0
sqlite=3.41.2=h5eee18b_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h04d1e81_0
urllib3=2.0.4=pyhd8ed1ab_0
wheel=0.38.4=py311h06a4308_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
zlib=1.2.13=h5eee18b_0
